expose final endpoint
