+++ services/api-server/src/simcore_service_api_server/api/routes/wallets.py
@@ -13,7 +13,9 @@ from simcore_service_api_server.api.dependencies.webserver_rpc import (
     get_wb_api_rpc_client,
 )
 from simcore_service_api_server.models.pagination import PaginationParams
-from simcore_service_api_server.models.schemas.wallets import LicensedItemCheckoutData
+from simcore_service_api_server.models.schemas.licensed_items import (
+    LicensedItemCheckoutData,
+)
 from simcore_service_api_server.services_rpc.wb_api_server import WbApiRpcClient
 
 from ...exceptions.service_errors_utils import DEFAULT_BACKEND_SERVICE_STATUS_CODES
@@ -114,26 +116,4 @@ async def checkout_licensed_item(
         licensed_item_id=licensed_item_id,
         num_of_seats=checkout_data.number_of_seats,
         service_run_id=checkout_data.service_run_id,
-    )
-
-
-@router.post(
-    "/{wallet_id}/licensed-items/{licensed_item_id}/release",
-    response_model=LicensedItemCheckoutGet,
-    status_code=status.HTTP_200_OK,
-    responses=WALLET_STATUS_CODES,
-    description="Release previously checked out licensed item",
-    include_in_schema=False,
-)
-async def release_licensed_item(
-    wallet_id: int,
-    licensed_item_id: LicensedItemID,
-    web_api_rpc: Annotated[WbApiRpcClient, Depends(get_wb_api_rpc_client)],
-    product_name: Annotated[str, Depends(get_product_name)],
-    user_id: Annotated[PositiveInt, Depends(get_current_user_id)],
-):
-    return await web_api_rpc.release_licensed_item_for_wallet(
-        product_name=product_name,
-        user_id=user_id,
-        licensed_item_checkout_id=licensed_item_id,
     )
