--- services/api-server/src/simcore_service_api_server/models/schemas/wallets.py
@@ -1,7 +1,0 @@
-from models_library.resource_tracker import ServiceRunId
-from pydantic import BaseModel
-
-
-class LicensedItemCheckoutData(BaseModel):
-    number_of_seats: int
-    service_run_id: ServiceRunId
